Python Script of the Project.
This is a part of the software for the project written in Python. This part of the software receives the keyword signals sent from the Arduino/C++ script through the COM3 port, and uses the PyAutoGUI module to automate the functions onto the Spotify app (i.e. previous track, pause track, next track).