Updated python script for song detection.
This is a part of the software for the project written in Python. This part of the software receives the keyword signals sent from the Arduino/C++ script through the COM3 port, and uses the PyAutoGUI module to automate the functions onto the Spotify app (i.e. previous track, pause track, next track). I updated this script to include a module for song detection called SwSpotify. The script can now send the song name to the Arduino through the COM3 port for display through the 1602 LCD.