Added song detection through 1602 LCD.
This is a part of the software for the project written in C++. This part of the software receives the crucial input from the buttons wired to the Arduino Uno. Based on the input received from the buttons, it prints a keyword onto the serial monitor which is communicated to the Python script through the COM3 port. This part of the script can now receive the song name from the python script and output it through the 1602 LCD.