Major Update: Added IR receiver functionality.
This is a major update to this project. This part of the software now receives the crucial input from an Infrared Remote and the input is processed using an Infrared Receiver. The buttons are no longer apart of the project as it is not user friendly. The IR receiver prints a keyword onto the serial monitor based on the key pressed on the transmitter (remote), which is communicated to the Python script through the COM3 port. Script can also retrieve the song name from the python script and output it through the 1602 LCD.